requirements.txt
# Backend Framework
fastapi==0.115.12
uvicorn[standard]==0.35.0
python-multipart==0.0.20
python-dotenv==1.0.1

# Document Processing
pandas==2.2.3
pypdf==5.1.0
unstructured==0.16.15

# LangChain Core
langchain==0.3.19
langchain-core==0.3.29
langchain-community==0.3.19
langchain-text-splitters==0.3.5
langchain-huggingface==1.2.0

# LLM Providers
langchain-groq==0.3.2
langchain-tavily==0.3.0
langchain-google-genai==2.0.11

# Embeddings
sentence-transformers==5.2.2

# Vector Database
pinecone==6.0.0

# Utilities
numpy==2.2.3
matplotlib==3.10.0
jinja2==3.1.5
